I FIXED IT.
Each EmptyCell holds a Cancer Cell arraylist which will contain a cancer cell created by Cancer class IF there parasitic condition is a Mycoplasma nearby.

Working on cleaning this code by not using ArrayList possibly and by increasing extendability so that fewer changes have to be made to classes outside of the one you're trying to make.