I have optimised the arraylistofspecialcells so that there is only one.
Index 0 is used for cancer cells and index 1 is used for mycoplasma cells. The arraylist is initialised with values 1 (which is why object type is Object so it can take integer values as well) so the arraylist is never empty. This way we don't have to have a check for whether it is empty, reducing the number of nested if statements.

This is also completely extendable. For any new kind of special cell, we simply initialise an extra value in the arraylist and use a new index in the arraylist instead of making a whole new arraylist.